tests/wpt/web-platform-tests/notifications/event-onshow.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Notifications Test: notification - onshow (basic)</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="author" title="Xin Liu" href="mailto:xinx.liu@intel.com">
<link rel="help" title="5 API" href="http://www.w3.org/TR/notifications/">
<meta name="flags" content="">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <div id="log"></div>
  <script>
var t = async_test("the notification.onshow event can be invoked", {timeout: 50000});

Notification.requestPermission(t.step_func(function(permission) {
    assert_equals(permission, "granted", "notification permission");

    var notification = new Notification("This is a notification.", {
        body: "Can you see this?."
    });
    notification.onshow = t.step_func(function(e) {
        assert_equals(Object.prototype.toString.call(e), "[object Event]", "the type of event");
        assert_equals(e.type, "show", "The event is the expected type.")
        t.done();
    });
}));
  </script>
</body>
